# Conflicts:
#	seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/checkpoint/operation/TaskAcknowledgeOperation.java
#	seatunnel-engine/seatunnel-engine-server/src/main/java/org/apache/seatunnel/engine/server/checkpoint/operation/TaskCompletedOperation.java

+++ seatunnel-engine/seatunnel-engine-server/src/test/java/org/apache/seatunnel/engine/server/TaskExecutionServiceTest.java
@@ -52,7 +52,7 @@ public class TaskExecutionServiceTest extends AbstractSeaTunnelServerTest {
     FlakeIdGenerator flakeIdGenerator;
     long taskRunTime = 2000;
     long jobId = 10001;
-    long pipeLineId = 100001;
+    int pipeLineId = 100001;
 
     @Before
     @Override
